docs: add props description
